requirement.txt
torch==2.6.0
opencv-python==4.11.0.86
Pillow==11.1.0
easyocr==1.7.2
ultralytics==8.3.87
